pyproject.toml
[build-system]
requires = ["setuptools>=77", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dnsguard"
version = "0.1.0"
description = "DNS resolution + security posture analyzer (lookup/trace/email/DNSSEC/report)"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "Mohamed Mohamud Ali" }]
dependencies = [
  "dnspython>=2.6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.5.0",
]

[project.scripts]
dnsguard = "dnsguard.cli:main"

[tool.setuptools]
include-package-data = true

# ✅ IMPORTANT: only package dnsguard, exclude data/ and other folders
[tool.setuptools.packages.find]
where = ["."]
include = ["dnsguard*"]
exclude = ["data*", "tests*", "docs*", ".github*"]

[tool.ruff]
line-length = 100
target-version = "py310"
